+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
+## [Unreleased]
+
+### Fixed
+
+* Fix Lua error when importing from spellbook [#201]
+
 ## [1.15.1] - 2024-07-25
 
 ### Added
@@ -18,7 +24,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 ### Fixed
 
 * Fix macro conditional sorting for `bonusbar`, `advflyable`, and dynamic flying
-* Fix lua error in the key visualizer when including action bar buttons [#199]
+* Fix Lua error in the key visualizer when including action bar buttons [#199]
 
 ### Changed
 
@@ -1478,6 +1484,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 [0.4.0]: https://github.com/Snakybo/Clicked/releases/tag/0.4.0
 [0.3.0]: https://github.com/Snakybo/Clicked/releases/tag/0.3.0
 
+[#201]: https://github.com/Snakybo/Clicked/issues/201
 [#199]: https://github.com/Snakybo/Clicked/issues/199
 [#193]: https://github.com/Snakybo/Clicked/issues/193
 [#191]: https://github.com/Snakybo/Clicked/issues/191